+++ MagicAPI/src/main/java/com/elmakers/mine/bukkit/api/batch/UndoBatch.java
@@ -1,4 +1,7 @@
 package com.elmakers.mine.bukkit.api.batch;
 
+import com.elmakers.mine.bukkit.api.block.UndoList;
+
 public interface UndoBatch extends Batch {
+    UndoList getUndoList();
 }