+++ Magic/CHANGELOG.md
@@ -30,6 +30,7 @@
  - Improvements to making sure data files don't get wiped
  - Use chunk ticket API for performance improvements in large construction and undo tasks in unloaded areas
  - Automaton will no longer force-load chunks
+ - Some fixes to undoing spells on world save, to prevent leftover spell changes
 
 # 7.9.12